% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_freq.R
\name{allele_count_fxn}
\alias{allele_count_fxn}
\title{Tabulate the frequency of each founder allele.}
\usage{
allele_count_fxn(z, founder_numbers)
}
\arguments{
\item{z}{numeric matrix containing haplotypes for one animal.}

\item{founder_numbers}{integers representing the founder alleles.}
}
\value{
numeric vector containing the frequency of each founder allele.
}
\description{
This function will count a founder allele even when the allele count is zero for one allele.
This is used to count the frequency of each founder letter in z.
}
