% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed.R
\name{breed}
\alias{breed}
\title{Given a population of anaimals, breed them for one generation.}
\usage{
breed(pop, markers)
}
\arguments{
\item{pop}{data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.}

\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}
}
\value{
a population of animals which have been bred.
}
\description{
We keep the order of the females constant and randomizes teh males before mating. We return a population of animals of the same size as the input population.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 4, markers)
}
