% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed_funnel.R
\name{breed_funnel}
\alias{breed_funnel}
\title{Breed a funnel of founders to a single mating pair.}
\usage{
breed_funnel(funnel, markers, nu = 11, p = 0.05)
}
\arguments{
\item{funnel}{list containing the founders set up to breed.}

\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}

\item{nu}{floating point number to pass to the xoi::simStahl function. Interference parameter.}

\item{p}{floating point number to pass to the xoi::simStahl function. Proportion of crossovers with no crossover interference.}
}
\value{
population list containing the haplotypes for a pair of animals.
}
\description{
This function accpets a breeding funnel and continues to call itself until there are only two animals, one female and one male, left.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
founders = 1:4
funnel = create_breeding_funnel(sample(founders), markers = markers)
mating_pair = breed_funnel(funnel, markers)
}
