% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed_one_x_chr.R
\name{breed_one_x_chr}
\alias{breed_one_x_chr}
\title{Breed the females and males on the X chromosome.}
\usage{
breed_one_x_chr(breeders, mkr, nu = 11, p = 0.05)
}
\arguments{
\item{breeders}{list containing the haplotype for a population of animals on one chromosome.}

\item{mkr}{named numeric vector containing the markers for the current chromosome.}

\item{nu}{floating point number to pass to the xoi::simStahl function. Interference parameter.}

\item{p}{floating point number to pass to the xoi::simStahl function. Proportion of crossovers with no crossover interference.}
}
\value{
list containing the haplotype for a population of animals on one chromosome.
}
\description{
This requires special handling since there is no crossing over in the males.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
founders = 1:4
breeders = create_breeding_funnel(hap_order = 1:4, markers, num_f1_pairs = 2)
one_chr = breed_one_x_chr(breeders[['X']], markers[['X']])
}
