% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_freq.R
\name{calc_allele_freq}
\alias{calc_allele_freq}
\title{Get the frequency of each founder allele on each chromosome.}
\usage{
calc_allele_freq(pop, founder_numbers)
}
\arguments{
\item{pop}{data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.}

\item{founder_numbers}{integers representing the founder alleles.}
}
\value{
list with one element per chromosome. Each element is a numeric matrix containing the proportion of each sample that is derived from each founder.
}
\description{
Get the frequency of each founder allele on each chromosome.
}
\examples{
n_mkr = 100
n_founders = 8
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = n_founders, markers)
allele_freq = calc_allele_freq(pop, 1:n_founders)
}
