% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_block_size.R
\name{calc_block_size}
\alias{calc_block_size}
\title{Calculate the recombination block size for each sample on each chromosome.}
\usage{
calc_block_size(pop, markers)
}
\arguments{
\item{pop}{data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.}

\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}
}
\value{
list containing recombination block sizes for each sample on each chromosome.
}
\description{
Calculate the recombination block size for each sample on each chromosome.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 8, markers)
blocks = calc_block_size(pop, markers)
}
