% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_crossovers.R
\name{count_crossovers}
\alias{count_crossovers}
\title{Count the number of crossovers in each sample on each chromosome.}
\usage{
count_crossovers(pop)
}
\arguments{
\item{pop}{data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.}
}
\value{
numeric matrix containing the number of crossovers for each sample on each chromosome.
}
\description{
Count the number of crossovers in each sample on each chromosome.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 8, markers)
xo = count_crossovers(pop)
}
