% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pop_list.R
\name{create_one_chr}
\alias{create_one_chr}
\title{Create data structure for one chromosome.}
\usage{
create_one_chr(mkr, n_pairs)
}
\arguments{
\item{mkr}{named numeric vector containing marker positions in cM and marker names.}

\item{n_pairs}{integer that is the number of pairs of breeding pairs to create.}
}
\value{
list contining two elements: females and males. Each list element is a named, three dimensional matrix with markers in rows, haplotypes in columns, and breeders in slices.
}
\description{
Create data structure for one chromosome.
}
\examples{
markers = lapply(1:2, function(z) { setNames(1:10, paste0('mkr', 1:10)) })
names(markers) = c(1, 'X')
pop4chr = create_one_chr(mkr = markers, n_pairs = 10)
}
