% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pop_list.R
\name{create_pop_list}
\alias{create_pop_list}
\title{Create a list to hold the population haplotypes.}
\usage{
create_pop_list(markers, n_pairs)
}
\arguments{
\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}

\item{n_pairs}{integer that is the number of pairs of breeders to create.}
}
\value{
list containing one element per chromosome. Each list element contains a list with two elements: female and male. The female and male elements are 3D integer arrays with markers in rows, haplotypes in columns, and sample in rows.
}
\description{
Create a list to hold the population haplotypes.
}
\examples{
markers = lapply(1:2, function(z) { setNames(1:10, paste0('mkr', 1:10)) })
names(markers) = c(1, 'X')
pop = create_pop_list(markers = markers, n_pairs = 10)
}
