% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample.R
\name{extract_sample}
\alias{extract_sample}
\title{Extract the haplotypes for one sample.}
\usage{
extract_sample(sex, index, pop)
}
\arguments{
\item{sex}{character string that is either 'feamles' or 'males'.}

\item{index}{index of the sample to extract.}

\item{pop}{data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.}
}
\value{
list containing one element per chromosome. Each element is a numeric matrix wtih markers in rows and two haplotypes in columns.
}
\description{
Extract the haplotypes for one sample.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 8, markers)
smpl = extract_sample('females', index = 5, pop)
}
