% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transmitted_chr.R
\name{get_transmitted_chr}
\alias{get_transmitted_chr}
\title{Get the transmitted chromosome from one meiosis.}
\usage{
get_transmitted_chr(haps, mkr, nu, p)
}
\arguments{
\item{haps}{integer matrix with markers in rows and two haplotypes in columns.}

\item{mkr}{named numeric vector containing the cM positions for each marker on the current chromosome.}

\item{nu}{floating point number to pass to the xoi::simStahl function. Interference parameter.}

\item{p}{floating point number to pass to the xoi::simStahl function. Proportion of crossovers with no crossover interference.}
}
\value{
numeric vector containing the sequence on one haplotype.
}
\description{
Given a pair of haplotypes for one animal, simulate crossovers and return a single haplotype, possibly with crossovers.
}
\examples{
n_mkr = 51
haps = matrix(rep(1:2, each = n_mkr), nrow = n_mkr, ncol = 2, 
dimnames = list(paste0('mkr', 1:n_mkr), paste0('hap', 1:2)))
mkr = setNames(1:n_mkr, paste0('mkr', 1:n_mkr))
trans_hap = get_transmitted_chr(haps, mkr, nu = 11, p = 0.05)
}
