% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot_genome}
\alias{plot_genome}
\title{Extract the haplotypes for one sample.}
\usage{
plot_genome(haps, markers, colors, ...)
}
\arguments{
\item{haps}{haplotypes for one sample from extract_sample.}

\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}

\item{colors}{vector of colors, one per founder.}

\item{...}{other argument to pass to plot.}
}
\value{
list containing one element per chromosome. Each element is a numeric matrix wtih markers in rows and two haplotypes in columns.
}
\description{
Extract the haplotypes for one sample.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 8, markers)
smpl = extract_sample('females', index = 5, pop)
plot_genome(smpl, markers, colors = 1:4)
}
