% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allele_freq.R
\name{reshape_allele_freq}
\alias{reshape_allele_freq}
\title{Reshape allele frequencies.}
\usage{
reshape_allele_freq(freq)
}
\arguments{
\item{freq}{list containing the allele frequencies generated by calc_allele_freq across multiple generations. Each list element contains the allele frequencies from one generation.}
}
\value{
list containing one element per chromsome. Each element is a three-dimensional, numeric array containing the allele frequencies. Markers in rows, founders in columns, and generations in slices.
}
\description{
Once the allele frequencies have been calculated, reshape them into three dimensional numerric arrays.
}
\examples{
n_mkr = 100
n_founders = 8
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = n_founders, markers)
allele_freq = as.list(1:2)
allele_freq[[1]] = calc_allele_freq(pop, 1:n_founders)
next_pop = breed(pop, markers)
allele_freq[[2]] = calc_allele_freq(next_pop, 1:n_founders)
freq = reshape_allele_freq(allele_freq)
}
