% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_founders.R
\name{simulate_founders}
\alias{simulate_founders}
\title{Simulate a first generation outcross.}
\usage{
simulate_founders(n_pairs = 175, n_founders = 8, markers)
}
\arguments{
\item{n_pairs}{integer that is the number of breeding pairs. Total animals will be 2 * n_pairs.}

\item{n_founders}{integer that is the number of founder animals. Must be a power of 2.}

\item{markers}{list containing one chromosome per element. Each element contains a named numeric vector containing marker positions in cM and marker names.}
}
\value{
data structure containing breeding population haplotypes. List containing one element per chromosome. Each element is a list with two elements named: females and males. Each element is a 3D numeric array with markers in rows, haplotypes in columns, and samples in slices.
}
\description{
Given a number of breeding pairs, number of founders, and the genetic markers, create a population of breeders derived directly from the breeding funnels.
}
\examples{
n_mkr = 100
markers = lapply(1:2, function(z) { setNames(1:n_mkr, paste0('mkr', 1:n_mkr)) })
names(markers) = c(1, 'X')
pop = simulate_founders(n_pairs = 10, n_founders = 8, markers)
}
